refactor(intraledger): move client calls to app
