+++ src/client/requests/mod.rs
@@ -1,8 +1,7 @@
 mod default_wallet;
 mod fetch_balance;
 mod globals;
-mod intraledger_send_usd;
-mod intraleger_send_btc;
+mod intraledger;
 mod me;
 mod set_username;
 mod user_login;
