+++ src/errors/mod.rs
@@ -1,12 +1,9 @@
 use thiserror::Error;
 
-use self::{
-    api_error::ApiError, me_error::MeError, payment_error::PaymentError, token_error::TokenError,
-};
+use self::{api_error::ApiError, me_error::MeError, token_error::TokenError};
 
 pub mod api_error;
 pub mod me_error;
-pub mod payment_error;
 pub mod token_error;
 
 #[derive(Error, Debug)]
@@ -17,6 +14,4 @@ pub enum CliError {
     ApiError(#[from] ApiError),
     #[error(transparent)]
     TokenError(#[from] TokenError),
-    #[error(transparent)]
-    PaymentError(#[from] PaymentError),
 }
