--- src/errors/payment_error.rs
@@ -1,7 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum PaymentError {
-    #[error("Failed to get {0} wallet")]
-    FailedToGetWallet(String),
-}
